Refactor bump helpers and docs
